In the class String, a change has been done in the operation plus().
Before a.plus(b) which is equivalent to a + b raised a CallOnVoidTarget when b was void.

Now b take the string value "<void>"

Note that a + b works now when b is void but still returns an exception when a is void. Then we have to use  "" + a + b.


git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/kermeta/trunk@3107 3b3ea266-f0f1-470e-95cb-5b2c9abc2e95

Former-commit-id: 2aa22d61fd2829225d19a785c30c700ba2633735