@pre improvment
To use @pre with a property obtained after navigation SEVERAL times

For example:

post thePost is
	theClass.parent@pre.name@pre != theClass.name

NOTE:
a case is still not consider when in a navigation an element has @pre we can not know the value of the precedent element of the navigation. For example "theClass" can not be evaluate before the execution of the method. A WARNING has been added.


git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/kermeta/trunk@3577 3b3ea266-f0f1-470e-95cb-5b2c9abc2e95

Former-commit-id: 99916ae4661be31fe2a681e196a67b8feed8e62a